(refactor) separates logic from events
