+++ IOMate.Api/Controllers/UserController.cs
@@ -1,9 +1,7 @@
-using IOMate.Application.Shared.Dtos;
 using IOMate.Application.UseCases.Users.CreateUser;
 using IOMate.Application.UseCases.Users.DeleteUser;
 using IOMate.Application.UseCases.Users.GetAllUsers;
 using IOMate.Application.UseCases.Users.GetUserById;
-using IOMate.Application.UseCases.Users.GetUserEvents;
 using IOMate.Application.UseCases.Users.UpdateUser;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
@@ -70,12 +68,5 @@ public class UsersController : ControllerBase
             return NotFound();
 
         return Ok(response);
-    }
-
-    [HttpGet("{id}/events")]
-    public async Task<ActionResult<List<GetEventResponseDto>>> GetUserEvents(Guid id, CancellationToken cancellationToken)
-    {
-        var result = await _mediator.Send(new GetUserEventsRequestDto(id), cancellationToken);
-        return Ok(result);
     }
 }
