+++ IOMate.Application/UseCases/Events/GetEvents/GetEventsRequestDto.cs
@@ -1,0 +1,6 @@
+using MediatR;
+
+namespace IOMate.Application.UseCases.Events.GetEvents
+{
+    public sealed record GetEventsRequestDto(Guid UserId) : IRequest<List<GetEventResponseDto>>;
+}
