--- IOMate.Application/UseCases/Users/GetUserEvents/GetUserEventsRequestDto.cs
@@ -1,7 +1,0 @@
-using IOMate.Application.Shared.Dtos;
-using MediatR;
-
-namespace IOMate.Application.UseCases.Users.GetUserEvents
-{
-    public sealed record GetUserEventsRequestDto(Guid UserId) : IRequest<List<GetEventResponseDto>>;
-}
